basic fix done!

+++ database/migrations/2020_02_29_192132_create_education_table.php
@@ -15,12 +15,12 @@ class CreateEducationTable extends Migration
     {
         Schema::create('education', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
             $table->foreign('user_id')
                   ->references('id')->on('users')
                   ->onDelete('cascade');
-                        
+
             $table->string('degree_name');
             $table->string('subject');
             $table->text('institute');
